+++ Examples/ConsumerSmoke/Sources/ConsumerSmoke/main.swift
@@ -1,6 +1,5 @@
 import Foundation
 import InnoNetwork
-import InnoNetworkProtobuf
 import InnoNetworkDownload
 import InnoNetworkWebSocket
 
@@ -25,7 +24,6 @@ let client = DefaultNetworkClient(
 )
 _ = client
 _ = ConsumerRequest()
-_ = ProtobufEmptyResponse.self
 _ = DownloadConfiguration.safeDefaults(sessionIdentifier: "com.example.consumer.downloads")
 _ = WebSocketConfiguration.safeDefaults()